Rename configs for mailer.
