+++ DIMS-Core/DIMS-Core.Mailer/mailersettings.json
